fix warnings

+++ lib/memory.hpp
@@ -154,7 +154,7 @@ class span
     size_t size() const restrict
     {
         assert(_end >= _begin);
-        return _end - _begin;
+        return size_t(_end - _begin);
     }
     bool empty() const restrict { return size() == 0; }
     operator bool() const restrict { return size() != 0; }
@@ -284,7 +284,7 @@ class unique_span : private D
         }
     }
 
-    size_t size() const { return _end - _begin; }
+    size_t size() const { return size_t(_end - _begin); }
     bool empty() const { return size() == 0; }
     operator bool() const { return size() != 0; }
 
@@ -421,8 +421,8 @@ alloc_mirrored(size_t size, size_t ncopies, const char* shm_name_base = nullptr)
     const size_t nbytes = sizeof(T) * size;
 
     // Create a backing file in tmpfs
-    unsigned fd = details::tmpshm(shm_name_base);
-    sys::check_ret(ftruncate(fd, nbytes), "ftrunctate");
+    int fd = details::tmpshm(shm_name_base);
+    sys::check_ret(ftruncate(fd, ssize_t(nbytes)), "ftrunctate");
 
     // Allocate the virtual contiguous address space
     T* ptr = reinterpret_cast<T*>(sys::check_ptr(mmap(nullptr, ncopies * nbytes, PROT_NONE, MAP_PRIVATE | MAP_ANONYMOUS, -1, 0), "mmap (anonymous)"));
